adjusts in CHCompTable
